mk_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
drf-spectacular==0.27.0
gunicorn==21.2.0
django-redis==5.4.0
twilio==8.10.0
africastalking==1.2.5
celery==5.3.4
django-ratelimit==4.1.0